requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.2.0
Django==3.2.20
django-crontab==0.7.1
django-filter==23.2
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
feedparser==6.0.10
idna==3.4
importlib-metadata==6.7.0
Markdown==3.4.3
pika==1.3.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sgmllib3k==1.0.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
zipp==3.15.0
